moved actions.js to actions folder
